+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Commit Message Formatter Changelog
 
+## [Upgrade Commit Message Formatter] - 2024-10-04
+- ✨ feat: support displaying multiple candidate commit messages on the search page
+
 ## [Upgrade Commit Message Formatter] - 2024-10-03
 - 🔧 chore: change project name to commit-message-formatter